+++ packages/airless-google-cloud-core/airless/google/cloud/core/operator/__init__.py
@@ -1,12 +1,10 @@
 from .base import (GoogleBaseEventOperator, GoogleBaseFileOperator)
 from .delay import (GoogleDelayOperator)
-from .error import (GoogleErrorReprocessOperator)
 from .redirect import (GoogleRedirectOperator)
 
 __all__ = [
     'GoogleBaseEventOperator',
     'GoogleBaseFileOperator',
     'GoogleDelayOperator',
-    'GoogleErrorReprocessOperator',
     'GoogleRedirectOperator'
 ]